strip_project/requirements.txt
Django==4.2.21
djangorestframework==3.14.0
stripe==7.11.0
python-dotenv==1.0.0
gunicorn==21.2.0 